+++ firebase-ai/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+- [changed] Added `LiveAudioConversationConfig` to control different aspects of the conversation
+  while using the `startAudioConversation` function.
+- [fixed] Fixed an issue causing streaming chat interactions to drop thought signatures. (#7562)
+
+# 17.6.0
+
 - [feature] Added support for server templates via `TemplateGenerativeModel` and
   `TemplateImagenModel`. (#7503)